+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/forms/resources/LpmRepository.kt
@@ -29,7 +29,6 @@ import com.stripe.android.paymentsheet.forms.GiropayRequirement
 import com.stripe.android.paymentsheet.forms.GrabPayRequirement
 import com.stripe.android.paymentsheet.forms.IdealRequirement
 import com.stripe.android.paymentsheet.forms.KlarnaRequirement
-import com.stripe.android.paymentsheet.forms.MandateRequirement
 import com.stripe.android.paymentsheet.forms.MobilePayRequirement
 import com.stripe.android.paymentsheet.forms.P24Requirement
 import com.stripe.android.paymentsheet.forms.PaymentMethodRequirements
@@ -52,6 +51,7 @@ import com.stripe.android.ui.core.elements.FormItemSpec
 import com.stripe.android.ui.core.elements.LayoutSpec
 import com.stripe.android.ui.core.elements.LpmSerializer
 import com.stripe.android.ui.core.elements.MandateTextSpec
+import com.stripe.android.ui.core.elements.NameSpec
 import com.stripe.android.ui.core.elements.SaveForFutureUseSpec
 import com.stripe.android.ui.core.elements.SepaMandateTextSpec
 import com.stripe.android.ui.core.elements.SharedDataSpec
@@ -250,7 +250,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Card.code -> SupportedPaymentMethod(
             code = "card",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_card,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_card,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -263,42 +262,27 @@ class LpmRepository constructor(
                 LayoutSpec(sharedDataSpec.fields)
             }
         )
-        PaymentMethod.Type.Bancontact.code -> {
-            val isSfu = (stripeIntent as? PaymentIntent)?.setupFutureUsage != null
-            val isSetupIntent = stripeIntent is SetupIntent
-            val localLayoutSpecs: List<FormItemSpec> = if (isSfu || isSetupIntent) {
-                listOf(
-                    EmailSpec(),
-                    SepaMandateTextSpec()
-                )
-            } else {
-                emptyList()
-            }
-
-            SupportedPaymentMethod(
-                code = "bancontact",
-                requiresMandate = true,
-                mandateRequirement = MandateRequirement.Always,
-                displayNameResource = R.string.stripe_paymentsheet_payment_method_bancontact,
-                iconResource = R.drawable.stripe_ic_paymentsheet_pm_bancontact,
-                lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
-                darkThemeIconUrl = sharedDataSpec.selectorIcon?.darkThemePng,
-                tintIconOnSelection = false,
-                requirement = BancontactRequirement,
-                formSpec = LayoutSpec(sharedDataSpec.fields + localLayoutSpecs)
-            )
-        }
+        PaymentMethod.Type.Bancontact.code -> SupportedPaymentMethod(
+            code = "bancontact",
+            requiresMandate = true,
+            displayNameResource = R.string.stripe_paymentsheet_payment_method_bancontact,
+            iconResource = R.drawable.stripe_ic_paymentsheet_pm_bancontact,
+            lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
+            darkThemeIconUrl = sharedDataSpec.selectorIcon?.darkThemePng,
+            tintIconOnSelection = false,
+            requirement = BancontactRequirement,
+            formSpec = LayoutSpec(sharedDataSpec.fields + stripeIntent.sepaMandateLayout())
+        )
         PaymentMethod.Type.Sofort.code -> SupportedPaymentMethod(
             code = "sofort",
             requiresMandate = true,
-            mandateRequirement = MandateRequirement.Always,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_sofort,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_klarna,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
             darkThemeIconUrl = sharedDataSpec.selectorIcon?.darkThemePng,
             tintIconOnSelection = false,
             requirement = SofortRequirement,
-            formSpec = LayoutSpec(sharedDataSpec.fields)
+            formSpec = LayoutSpec(sharedDataSpec.fields + stripeIntent.sepaMandateLayout(requireName = true))
         )
         PaymentMethod.Type.Ideal.code -> {
             val isSfu = (stripeIntent as? PaymentIntent)?.setupFutureUsage != null
@@ -315,7 +299,6 @@ class LpmRepository constructor(
             SupportedPaymentMethod(
                 code = "ideal",
                 requiresMandate = true,
-                mandateRequirement = MandateRequirement.Always,
                 displayNameResource = R.string.stripe_paymentsheet_payment_method_ideal,
                 iconResource = R.drawable.stripe_ic_paymentsheet_pm_ideal,
                 lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -328,7 +311,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.SepaDebit.code -> SupportedPaymentMethod(
             code = "sepa_debit",
             requiresMandate = true,
-            mandateRequirement = MandateRequirement.Always,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_sepa_debit,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_sepa_debit,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -340,7 +322,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Eps.code -> SupportedPaymentMethod(
             code = "eps",
             requiresMandate = true,
-            mandateRequirement = MandateRequirement.Always,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_eps,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_eps,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -352,7 +333,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.P24.code -> SupportedPaymentMethod(
             code = "p24",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_p24,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_p24,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -364,7 +344,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Giropay.code -> SupportedPaymentMethod(
             code = "giropay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_giropay,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_giropay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -376,7 +355,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.AfterpayClearpay.code -> SupportedPaymentMethod(
             code = "afterpay_clearpay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = if (isClearpay()) {
                 R.string.stripe_paymentsheet_payment_method_clearpay
             } else {
@@ -392,7 +370,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Klarna.code -> SupportedPaymentMethod(
             code = "klarna",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_klarna,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_klarna,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -411,7 +388,6 @@ class LpmRepository constructor(
             SupportedPaymentMethod(
                 code = "paypal",
                 requiresMandate = stripeIntent.payPalRequiresMandate(),
-                mandateRequirement = MandateRequirement.Dynamic,
                 displayNameResource = R.string.stripe_paymentsheet_payment_method_paypal,
                 iconResource = R.drawable.stripe_ic_paymentsheet_pm_paypal,
                 lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -424,7 +400,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Affirm.code -> SupportedPaymentMethod(
             code = "affirm",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_affirm,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_affirm,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -436,7 +411,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.RevolutPay.code -> SupportedPaymentMethod(
             code = "revolut_pay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_revolut_pay,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_revolut_pay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -448,7 +422,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.AmazonPay.code -> SupportedPaymentMethod(
             code = "amazon_pay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_amazon_pay,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_amazon_pay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -460,7 +433,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.MobilePay.code -> SupportedPaymentMethod(
             code = "mobilepay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_mobile_pay,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_mobile_pay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -472,7 +444,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Zip.code -> SupportedPaymentMethod(
             code = "zip",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_zip,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_zip,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -484,7 +455,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.AuBecsDebit.code -> SupportedPaymentMethod(
             code = "au_becs_debit",
             requiresMandate = true,
-            mandateRequirement = MandateRequirement.Always,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_au_becs_debit,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_bank,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -514,7 +484,6 @@ class LpmRepository constructor(
                 SupportedPaymentMethod(
                     code = "us_bank_account",
                     requiresMandate = true,
-                    mandateRequirement = MandateRequirement.Always,
                     displayNameResource = R.string.stripe_paymentsheet_payment_method_us_bank_account,
                     iconResource = R.drawable.stripe_ic_paymentsheet_pm_bank,
                     lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -530,7 +499,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Upi.code -> SupportedPaymentMethod(
             code = "upi",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_upi,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_upi,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -543,7 +511,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Blik.code -> SupportedPaymentMethod(
             code = "blik",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_blik,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_blik,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -556,7 +523,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.CashAppPay.code -> SupportedPaymentMethod(
             code = "cashapp",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_cashapp,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_cash_app_pay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -568,7 +534,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.GrabPay.code -> SupportedPaymentMethod(
             code = "grabpay",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_grabpay,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_grabpay,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -580,7 +545,6 @@ class LpmRepository constructor(
         PaymentMethod.Type.Fpx.code -> SupportedPaymentMethod(
             code = "fpx",
             requiresMandate = false,
-            mandateRequirement = MandateRequirement.Never,
             displayNameResource = R.string.stripe_paymentsheet_payment_method_fpx,
             iconResource = R.drawable.stripe_ic_paymentsheet_pm_fpx,
             lightThemeIconUrl = sharedDataSpec.selectorIcon?.lightThemePng,
@@ -621,9 +585,6 @@ class LpmRepository constructor(
 
         /** This describes if the LPM requires a mandate see [ConfirmPaymentIntentParams.mandateDataParams]. */
         val requiresMandate: Boolean,
-
-        /** This describes when the LPM requires a mandate (see [ConfirmPaymentIntentParams.mandateDataParams]). */
-        val mandateRequirement: MandateRequirement,
 
         /** This describes the name that appears under the selector. */
         @StringRes val displayNameResource: Int,
@@ -718,7 +679,6 @@ class LpmRepository constructor(
             return SupportedPaymentMethod(
                 code = "card",
                 requiresMandate = false,
-                mandateRequirement = MandateRequirement.Never,
                 displayNameResource = R.string.stripe_paymentsheet_payment_method_card,
                 iconResource = R.drawable.stripe_ic_paymentsheet_pm_card,
                 lightThemeIconUrl = null,
@@ -750,5 +710,25 @@ private fun StripeIntent.payPalRequiresMandate(): Boolean {
     return when (this) {
         is PaymentIntent -> setupFutureUsage != null
         is SetupIntent -> true
+    }
+}
+
+private fun StripeIntent.sepaMandateLayout(
+    requireName: Boolean = false,
+): List<FormItemSpec> {
+    val isSfu = (this as? PaymentIntent)?.setupFutureUsage != null
+    val isSetupIntent = this is SetupIntent
+    return if (isSfu || isSetupIntent) {
+        listOfNotNull(
+            if (requireName) {
+                NameSpec()
+            } else {
+                null
+            },
+            EmailSpec(),
+            SepaMandateTextSpec()
+        )
+    } else {
+        emptyList()
     }
 }