+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestAuBecsDD.kt
@@ -2,8 +2,10 @@ package com.stripe.android.lpm
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.stripe.android.BaseLpmTest
+import com.stripe.android.test.core.Automatic
 import com.stripe.android.test.core.Currency
 import com.stripe.android.test.core.DelayedPMs
+import com.stripe.android.test.core.IntentType
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -28,6 +30,28 @@ internal class TestAuBecsDD : BaseLpmTest() {
     fun testAuBecsDDInCustomFlow() {
         testDriver.confirmCustom(
             testParameters = auBecsDD,
+        )
+    }
+
+    @Test
+    fun testAuBecsDDSfu() {
+        testDriver.confirmNewOrGuestComplete(
+            testParameters = auBecsDD.copy(
+                delayed = DelayedPMs.On,
+                automatic = Automatic.On,
+                intentType = IntentType.PayWithSetup,
+            ),
+        )
+    }
+
+    @Test
+    fun testAuBecsDDSetup() {
+        testDriver.confirmNewOrGuestComplete(
+            testParameters = auBecsDD.copy(
+                delayed = DelayedPMs.On,
+                automatic = Automatic.On,
+                intentType = IntentType.Setup,
+            ),
         )
     }
 }