+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### PaymentSheet
 
 * [ADDED][7269](https://github.com/stripe/stripe-android/pull/7269) PaymentSheet now supports Ideal with SetupIntent and PaymentIntent with setup for future usage.
+* [ADDED][7270](https://github.com/stripe/stripe-android/pull/7270) PaymentSheet now supports SEPA with SetupIntent and PaymentIntent with setup for future usage.
+* [ADDED][7272](https://github.com/stripe/stripe-android/pull/7272) PaymentSheet now supports Sofort with SetupIntent and PaymentIntent with setup for future usage.
+* [ADDED][7273](https://github.com/stripe/stripe-android/pull/7273) PaymentSheet now supports BECS Direct Debit with SetupIntent and PaymentIntent with setup for future usage.
 * [ADDED][7274](https://github.com/stripe/stripe-android/pull/7274) PaymentSheet now supports Alipay for PaymentIntents.
 
 ## 20.29.2 - 2023-09-05